split jit modules
